+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Add comment to shared note/link
+* Add copy buttons for restoring archived projects
 
 
 ## v2025.90 - 2025-10-22